estan las unidades del ingrediente

+++ Models/Plato.cs
@@ -3,7 +3,7 @@ public class Plato
     public int id { get; set; }
     public string nombre { get; set; }
     public int idCategoria { get; set; }
-    public int precio { get; set; }
+    public double precio { get; set; }
     public bool activo { get; set; }
 
 }
